backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.0
PyJWT==2.8.0
mysql-connector-python==8.3.0